Refactor inventory models into modules
