+++ inventory/models/suppliers.py
@@ -1,0 +1,19 @@
+from django.db import models
+
+
+class Supplier(models.Model):
+    supplier_id = models.AutoField(primary_key=True)
+    name = models.CharField(max_length=255, unique=True, null=False, blank=False)
+    contact_person = models.CharField(max_length=255, blank=True, null=True)
+    phone = models.CharField(max_length=20, blank=True, null=True)
+    email = models.CharField(max_length=254, blank=True, null=True)
+    address = models.TextField(blank=True, null=True)
+    notes = models.TextField(blank=True, null=True)
+    is_active = models.BooleanField(default=False, null=False)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self) -> str:  # pragma: no cover - simple representation
+        return self.name or f"Supplier {self.pk}"
+
+    class Meta:
+        db_table = "suppliers"
